Added new admin layout

+++ app/Controllers/Http/PostController.js
@@ -1,5 +1,3 @@
-'use strict'
-
 const { UserNotFoundException } = require('@adonisjs/auth/src/Exceptions')
 const { RuntimeException } = require('@adonisjs/lucid/src/Exceptions')
 
@@ -13,8 +11,9 @@ const Post = use('App/Models/Post')
 const User = use('App/Models/User')
 /** @type {typeof import('@adonisjs/../../app/Models/Comment')} */
 const Comment = use('App/Models/Comment')
+const NotFoundException = use('App/Exceptions/NotFoundException')
 
-const CloudinaryService = use('App/Services/CloudinaryService');
+const cloudinary = use('App/Services/CloudinaryService');
 const Helpers = use('Helpers')
 
 /**
@@ -30,10 +29,20 @@ class PostController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async index ({ request, response, view }) {
-    const posts = await Post.query().with('comments').with('tags').with('likes').with('user').fetch()
+  async index({ request, response, view }) {
+    const posts = await Post.query()
+      .with('comments')
+      .with('tags')
+      .with('likes')
+      .with('user')
+      .orderBy('created_at')
+      .fetch()
+
+    // const admin = await User.findBy('role', 'admin')
+    // console.log(posts.rows);
     return view.render('index', {
       posts: posts.toJSON(),
+      // admin: admin.toJSON(),
       meta: request.meta
     })
   }
@@ -47,8 +56,8 @@ class PostController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async create ({ request, response, view }) {
-    return view.render('post_add');
+  async create({ view }) {
+    return view.render('posts/add');
   }
 
   /**
@@ -59,34 +68,37 @@ class PostController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async store ({ request, response, session, auth }) {
-    const {title, text} = request.all()
+  async store({ request, response, session, auth }) {
+    const { title, text } = request.all()
     const img = request.file('image', {
       types: ['image'],
       size: '2mb'
     })
-    try{
-      const post = new Post()
-        // await img.move(Helpers.tmpPath('uploads'), {
-        // name: `${new Date().getTime()}.${file.subtype}`
-        // })
-        // if (!img.moved()) {
-        //   session.flash({error: img.error()})
-        //   return response.redirect('back')
-        // }
-        // post.img_path = img.fileName
-      if(img) {
-        const cloudinaryResponse = await CloudinaryService.v2.uploader.upload(img.tmpPath, {folder: 'forum/uploads'});
-        post.img_path = cloudinaryResponse.secure_url
-      }
-      post.title = title
-      post.text = text
-      post.user_id = auth.user.id
-      await post.save()
-      session.flash({success: 'Sucessfully created a post'})
-    } catch(e) {
-      session.flash({error: e.message})
+
+    const post = new Post()
+    // await img.move(Helpers.tmpPath('uploads'), {
+    // name: `${new Date().getTime()}.${file.subtype}`
+    // })
+    // if (!img.moved()) {
+    //   session.flash({error: img.error()})
+    //   return response.redirect('back')
+    // }
+    // post.img_path = img.fileName
+    if (img) {
+      const cloudinaryResponse = await cloudinary.v2.uploader.upload(
+        img.tmpPath,
+        {
+          folder: 'forum/uploads', width: 500, height: 333, crop: 'fill'
+        }
+      )
+      post.img_path = cloudinaryResponse.secure_url
     }
+    post.title = title
+    post.text = text
+    post.user_id = auth.user.id
+    await post.save()
+    session.flash({ success: 'Sucessfully created a post' })
+
     return response.redirect('/');
   }
 
@@ -99,11 +111,27 @@ class PostController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async show ({ params, request, response, view }) {
-    const {id} = params
-    const post = await Post.query().where('id', id).with('user').first()
-    post.comments = (await Comment.query().where('post_id', post.id).with('user').fetch()).toJSON()
-    return view.render('post', {post: post.toJSON(), meta: request.meta})
+  async show({ params, view, request, auth }) {
+    const { id } = params
+    const post = await Post
+      .query()
+      .where('id', id)
+      .with('user')
+      .with('likes')
+      .first()
+    if(!post) throw new NotFoundException()
+    post.comments = (await Comment
+      .query()
+      .where('post_id', post.id)
+      .with('user')
+      .fetch()).toJSON()
+    post.is_liked = post.$relations.likes.rows.filter(like => 
+      like.user_id === auth.user.id
+    ).length > 0
+    return view.render('posts.post', { 
+      post: post.toJSON(),
+      meta: request.meta
+    })
   }
 
   /**
@@ -115,36 +143,38 @@ class PostController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-   async search ({ params, request, response, view }) {
-    const {tag, title, time, user} = request.all()
-    var queryToDisplay = ""; 
-    var query = Post.query();
-    if(tag) {
-      query = query.whereHas('tags', (builder) => {
+  async search({ request, view }) {
+    const {
+      tag, title, time, user
+    } = request.all()
+    let queryToDisplay = [];
+    let query = Post.query();
+    if (tag) {
+      query.whereHas('tags', (builder) => {
         builder.where('value', tag)
       })
-      queryToDisplay += `tag: ${tag}`
+      queryToDisplay.push(tag);
     }
-    if(title) {
+    if (title) {
       query = query.where('title', 'like', `%${title}%`)
-      if(queryToDisplay)
-        queryToDisplay += `, `
-      queryToDisplay += `title: ${title}`
+      queryToDisplay.push(title);
     }
-    if(time) {
+    if (time) {
       query = query.where(`datediff(day, created_at, ${time})`, 0)
-      if(queryToDisplay)
-        queryToDisplay += `, `
-      queryToDisplay += `time: ${time}`
+      queryToDisplay.push(time)
     }
-    if(user) {
-      query = query.where(`user_id`, user)
-      if(queryToDisplay)
-        queryToDisplay += `, `
-      queryToDisplay += `user: ${(await User.find(user)).username}`
+    if (user) {
+      query = query.where('user_id', user)
+      queryToDisplay.push((await User.find(user)).username)
     }
-    const posts = await query.with('user').with('likes').with('comments').fetch()
-    return view.render('search', {
+
+    const posts = await query
+      .with('user')
+      .with('likes')
+      .with('comments')
+      .fetch()
+
+    return view.render('posts/search', {
       posts: posts.toJSON(),
       meta: request.meta,
       query: queryToDisplay
@@ -160,10 +190,10 @@ class PostController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async edit ({ params, request, response, view }) {
-    const {id} = params
+  async edit({ params, request, response, view }) {
+    const { id } = params
     const post = await Post.find(id)
-    return view.render('post_edit', {post: post.toJSON()})
+    return view.render('posts/edit', { post: post.toJSON() })
   }
 
   /**
@@ -174,29 +204,35 @@ class PostController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async update ({ params, request, response, session }) {
-    const {id} = params
-    const {title, text} = request.all()
+  async update({ params, request, response, session }) {
+    const { id } = params
+    const { title, text } = request.all()
     const img = request.file('img', {
       types: ['image'],
       size: '2mb',
       extnames: ['png', 'jpg', 'jpeg']
     })
     const post = await Post.find(id)
-    try{
-      if(title) post.title = title
-      if(text) post.text = text
-      if(img) {
-        await img.move(Helpers.tmpPath('uploads'), {
-          name: `${new Date().getTime()}.${file.subtype}`
-        })
-        if (!img.moved()) {
-          throw new RuntimeException
+    if (title) post.title = title
+    if (text) post.text = text
+    if (img) {
+      // await img.move(Helpers.tmpPath('uploads'), {
+      //   name: `${new Date().getTime()}.${file.subtype}`
+      // })
+      // if (!img.moved()) {
+      //   throw new RuntimeException
+      // }
+      const imgFileName = post.img_path.split('/').at(-1)
+      const imgId = `forum/uploads/${imgFileName.slice(0, imgFileName.indexOf('.'))}`
+      await cloudinary.v2.uploader.destroy(imgId, { invalidate: true, resource_type: 'image' })
+
+      const cloudinaryResponse = await cloudinary.v2.uploader.upload(
+        img.tmpPath,
+        {
+          folder: 'forum/uploads', width: 500, height: 333, crop: 'fill'
         }
-        post.img_path = img.fileName
-      }
-    } catch(e) {
-      session.flash({error: 'Error updating the post'})
+      )
+      post.img_path = cloudinaryResponse.secure_url
     }
     return response.redirect('back')
   }
@@ -209,9 +245,13 @@ class PostController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async destroy ({ params, request, response }) {
-    const {id} = params;
+  async destroy({ params, response }) {
+    const { id } = params;
     const post = await Post.find(id)
+    if (!post) throw new NotFoundException()
+    const imgFileName = post.img_path.split('/').at(-1)
+    const imgId = `forum/uploads/${imgFileName.slice(0, imgFileName.indexOf('.'))}`
+    await cloudinary.v2.uploader.destroy(imgId, { invalidate: true, resource_type: 'image' })
     await post.delete()
     return response.redirect('back')
   }
